Some clean up was done
