+++ twitter_clone/templates/own_feed.html
@@ -41,7 +41,7 @@
         <div class="row">
             <form role="form" action="" method="POST">
     		  <div class="form-group">
-                <textarea class="form-control" name="tweet" placeholder="Write your tweet here..." required></textarea>
+                <textarea class="form-control" name="tweet" maxlength="140" placeholder="Write your tweet here..." required></textarea>
     		  </div>
     		  <button type="submit" class="btn btn-info pull-right">Submit</button>
     		</form>
